refactor(fractional indexing): split fractional index structure into more focused files
