+++ server/utils/fractional-indexing/index.ts
@@ -1,0 +1,13 @@
+export * from './types.js';
+export { DEFAULT_INCREMENT } from './calculations.js';
+export { 
+	calculateMidpoint, 
+	calculateAfter, 
+	generateSequence 
+} from './calculations.js';
+export { 
+	getNextAppendIndex, 
+	moveWithFractionalIndex, 
+	bulkMoveToGearCloset, 
+	rebalanceIndexes 
+} from './database-operations.js';
